+++ diagrams/__init__.py
@@ -96,9 +96,8 @@ class _Cluster:
 
     def _validate_direction(self, direction: str):
         direction = direction.upper()
-        for v in self.__directions:
-            if v == direction:
-                return True
+        if direction in self.__directions:
+            return True
         return False
 
     def __str__(self) -> str:
@@ -107,7 +106,7 @@ class _Cluster:
 
 class Diagram(_Cluster):
     __curvestyles = ("ortho", "curved")
-    __outformats = ("png", "jpg", "svg", "pdf")
+    __outformats = ("png", "jpg", "svg", "pdf", "dot")
 
     # fmt: off
     _default_graph_attrs = {
@@ -198,8 +197,13 @@ class Diagram(_Cluster):
             raise ValueError(f'"{curvestyle}" is not a valid curvestyle')
         self.dot.graph_attr["splines"] = curvestyle
 
-        if not self._validate_outformat(outformat):
-            raise ValueError(f'"{outformat}" is not a valid output format')
+        if isinstance(outformat, list):
+            for one_format in outformat:
+                if not self._validate_outformat(one_format):
+                    raise ValueError(f'"{one_format}" is not a valid output format')
+        else:
+            if not self._validate_outformat(outformat):
+                raise ValueError(f'"{outformat}" is not a valid output format')
         self.outformat = outformat
 
         # Merge passed in attributes
@@ -238,16 +242,14 @@ class Diagram(_Cluster):
 
     def _validate_curvestyle(self, curvestyle: str) -> bool:
         curvestyle = curvestyle.lower()
-        for v in self.__curvestyles:
-            if v == curvestyle:
-                return True
+        if curvestyle in self.__curvestyles:
+            return True
         return False
 
     def _validate_outformat(self, outformat: str) -> bool:
         outformat = outformat.lower()
-        for v in self.__outformats:
-            if v == outformat:
-                return True
+        if outformat in self.__outformats:
+            return True
         return False
 
     def connect(self, node: "Node", node2: "Node", edge: "Edge") -> None:
@@ -255,7 +257,11 @@ class Diagram(_Cluster):
         self.edges[(node, node2)] = edge
 
     def render(self) -> None:
-        self.dot.render(format=self.outformat, view=self.show, quiet=True)
+        if isinstance(self.outformat, list):
+            for one_format in self.outformat:
+                self.dot.render(format=one_format, view=self.show, quiet=True)
+        else:
+            self.dot.render(format=self.outformat, view=self.show, quiet=True)
 
 
 class Node(_Cluster):